Conflicts:
	code/DriveIT/DriveIT.WindowsClient/ViewModels/CustomerViewModel.cs FIXED

+++ code/DriveIT/DriveIT.CarQuery/CarQuery.cs
@@ -39,14 +39,14 @@ namespace DriveIT.CarQuery
             {
                 if (result == null) result = await JSONWrapper.Read<TrimArray>(query);
                 if (result.Trims[0].model_0_to_100_kph.HasValue) 
-                    carToFill.NoughtTo100 = result.Trims[0].model_0_to_100_kph.Value;
+                    carToFill.NoughtTo100 = (float)result.Trims[0].model_0_to_100_kph.Value;
             }
 
             if (carToFill.TopSpeed == 0)
             {
                 if (result == null) result = await JSONWrapper.Read<TrimArray>(query);
                 if (result.Trims[0].model_top_speed_kph.HasValue) 
-                    carToFill.TopSpeed = result.Trims[0].model_top_speed_kph.Value;
+                    carToFill.TopSpeed = (float)result.Trims[0].model_top_speed_kph.Value;
             }
 
             if (string.IsNullOrEmpty(carToFill.Transmission) || string.IsNullOrWhiteSpace(carToFill.Transmission))
